requirements.txt
fastapi==0.115.0
pyserial==3.5
pytest==8.3.3
uvicorn==0.30.6
